+++ tests/test_text_to_image_models.py
@@ -1,6 +1,7 @@
 import os
 import requests
 import sys
+from requests.models import Response
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.append(os.path.join(os.path.dirname(__file__), '../llmmaster'))
@@ -116,7 +117,8 @@ def test_stable_diffusion_text_to_image_instances(run_api):
         for name, response in master.results.items():
             if isinstance(response, str):
                 print(f'{name} = {response}')
-            else:
+
+            elif isinstance(response, Response):
                 if 'output_format' in master.instances[name].parameters:
                     save_as = f'{name}.{master.instances[name].parameters["output_format"]}'
                 else:
@@ -124,11 +126,14 @@ def test_stable_diffusion_text_to_image_instances(run_api):
 
                 filepath = os.path.join(TEST_OUTPUT_PATH, save_as)
                 with open(filepath, 'wb') as f:
-                    if isinstance(response, bytes):
-                        f.write(response)
+                    if isinstance(response.content, bytes):
+                        f.write(response.content)
                         print(f'{name} = {filepath} saved')
                     else:
-                        print(f"Unexpected response type for {name}: {type(response)}")
+                        print(f"Unexpected response type for response.content of {name}: {type(response.content)}")
+
+            else:
+                pytest.fail(f'Unexpected response type for {name}: {type(response)}')
 
     print(f'Elapsed time (sec): {master.elapsed_time}')
     master.dismiss()